+++ http/src/test/java/com/twitter/finatra/http/tests/integration/doeverything/main/DoEverythingJavaController.java
@@ -5,7 +5,9 @@ import javax.inject.Inject;
 import scala.runtime.AbstractFunction0;
 import scala.runtime.BoxedUnit;
 
+import com.twitter.finagle.http.Method;
 import com.twitter.finagle.http.Request;
+import com.twitter.finagle.http.RouteIndex;
 import com.twitter.finatra.http.AbstractController;
 import com.twitter.util.Future;
 
@@ -26,14 +28,55 @@ public class DoEverythingJavaController extends AbstractController {
 
         get("/goodbye", (Request request) -> new GoodbyeResponse("guest", "cya", 123));
 
+        get(
+            "/bar/baz/foo",
+            "foo_get",
+            false,
+            null,
+            (Request request) -> new GoodbyeResponse("guest", "cya", 123));
+
+        get("/admin/foo", true, (Request request) -> "admin route no index");
+
+        get("/admin/finatra/implied", (Request request) -> "implied admin route no index");
+
+        get(
+            "/admin/clients",
+            true,
+            RouteIndex.apply(
+                "Special",
+                "Finatra",
+                scala.Option.apply(null),
+                Method.Get$.MODULE$),
+            (Request request) -> "admin route with index");
+
         get("/query", request ->
             helloService.computeQueryResult(request.getParam("q")));
 
         post("/post",  (Request request) -> "post");
 
+        post(
+            "/admin/bar",
+            true,
+            RouteIndex.apply(
+                "Special",
+                "Finatra",
+                scala.Option.apply(null),
+                Method.Post$.MODULE$),
+            (Request request) -> "post");
+
         put("/put",  (Request request) -> "put");
 
         delete("/delete",  (Request request) -> "delete");
+
+        delete(
+            "/admin/delete",
+            true,
+            RouteIndex.apply(
+                    "Special",
+                    "Finatra",
+                    scala.Option.apply(null),
+                    Method.Post$.MODULE$),
+            (Request request) -> "delete");
 
         options("/options",  (Request request) -> "options");